# context/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "details": {
            "description": "Detailed validation errors",
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "error": {
            "description": "Stable error category",
            "type": "string"
          },
          "message": {
            "description": "Human-readable error detail",
            "type": "string"
          }
        },
        "type": "object"
      },
      "User": {
        "properties": {
          "baseCurrency": {
            "description": "User portfolio base currency (3 letters)",
            "example": "USD",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "description": "User email address",
            "format": "email",
            "type": "string"
          },
          "id": {
            "description": "Unique identifier",
            "type": "string"
          },
          "name": {
            "description": "User full name",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Enter the access token returned by /api/auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "API Support",
      "url": "https://github.com/bruno-frascino/piggy-api"
    },
    "description": "A comprehensive stock portfolio management API built with Express and Prisma",
    "title": "Truffles API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "delete": {
        "description": "Marks the account deleted and revokes every session immediately. The data is retained until the grace period expires, during which signing in returns 403 with `accountPendingDeletion` and the account can be restored via POST /api/auth/restore. After that a scheduled purge removes the user and all their positions, accounts, snapshots, tax reports and watchlists.\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  }
                },
                "required": [
                  "currentPassword"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "properties": {
                        "deletedAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "purgeAfter": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Account scheduled for deletion"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Current password incorrect or not authenticated"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Schedule the authenticated user's account for deletion",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "User profile"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get the authenticated user's profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "baseCurrency": {
                    "example": "USD",
                    "maxLength": 3,
                    "minLength": 3,
                    "type": "string"
                  },
                  "currentPassword": {
                    "description": "Required when changing password",
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "newPassword": {
                    "description": "Must be at least 8 characters",
                    "minLength": 8,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Profile updated"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "currentPassword is required when setting a new password"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Current password incorrect or not authenticated"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update the authenticated user's profile or password",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:4000"
    }
  ],
...
}